# openapi.json
{"openapi":"3.1.0","info":{"title":"MediRAG API","description":"Clinical RAG system providing grounded evidence-based answers.","version":"v1.0 (MVP)"},"paths":{"/api/v1/auth/register":{"post":{"tags":["auth"],"summary":"Register","operationId":"register_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegister"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_v1_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Token"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["auth"],"summary":"Read Users Me","operationId":"read_users_me_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserOut"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/documents/upload":{"post":{"tags":["documents"],"summary":"Upload Document","operationId":"upload_document_api_v1_documents_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_document_api_v1_documents_upload_post"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DocumentOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/documents/":{"get":{"tags":["documents"],"summary":"List Documents","description":"Lists all uploaded medical documents in the system.","operationId":"list_documents_api_v1_documents__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DocumentOut"},"type":"array","title":"Response List Documents Api V1 Documents  Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/documents/{doc_id}":{"delete":{"tags":["documents"],"summary":"Delete Document","description":"Deletes a document and its corresponding vector embeddings.","operationId":"delete_document_api_v1_documents__doc_id__delete","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"doc_id","in":"path","required":true,"schema":{"type":"integer","title":"Doc Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/documents/status/{doc_id}":{"get":{"tags":["documents"],"summary":"Document Status","description":"Returns processing status and chunk count for a given document id.","operationId":"document_status_api_v1_documents_status__doc_id__get","security":[{"OAuth2PasswordBearer":[]}],"parameters":[{"name":"doc_id","in":"path","required":true,"schema":{"type":"integer","title":"Doc Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/chat/query":{"post":{"tags":["chat"],"summary":"Clinical Query","description":"RAG Endpoint:\n1. Embeds query.\n2. Retrieves top-k relevant document chunks.\n3. Prompts LLM to write a grounded response with page citations.\n4. Saves search query to history.","operationId":"clinical_query_api_v1_chat_query_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QueryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/chat/feedback":{"post":{"tags":["chat"],"summary":"Submit Feedback","description":"Allows users to submit thumbs up/down ratings and optional comments.","operationId":"submit_feedback_api_v1_chat_feedback_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FeedbackOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/chat/history":{"get":{"tags":["chat"],"summary":"Get Query History","description":"Fetches full conversational query logs and rating feedbacks for the authenticated user.","operationId":"get_query_history_api_v1_chat_history_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/QueryHistoryOut"},"type":"array","title":"Response Get Query History Api V1 Chat History Get"}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/admin/llm/validate":{"get":{"tags":["admin"],"summary":"Validate Llm","description":"Validate LLM/Embedding service connectivity. Admin-only recommended.\n\nReturns a lightweight summary indicating whether the Gemini client appears\nconnected and a small preview of available models when possible.","operationId":"validate_llm_api_v1_admin_llm_validate_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/admin/llm/reload":{"post":{"tags":["admin"],"summary":"Reload Llm","description":"Reload LLM service configuration (reads `.env` and reinitializes client).","operationId":"reload_llm_api_v1_admin_llm_reload_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/api/v1/admin/llm/debug_models":{"get":{"tags":["admin"],"summary":"Debug Models","description":"Return a small serialized preview of the raw models.list() response for debugging.","operationId":"debug_models_api_v1_admin_llm_debug_models_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"OAuth2PasswordBearer":[]}]}},"/":{"get":{"summary":"Root","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"Body_login_api_v1_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"password"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_api_v1_auth_login_post"},"Body_upload_document_api_v1_documents_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"title":{"type":"string","title":"Title"}},"type":"object","required":["file","title"],"title":"Body_upload_document_api_v1_documents_upload_post"},"CitationInfo":{"properties":{"document_id":{"type":"integer","title":"Document Id"},"document_title":{"type":"string","title":"Document Title"},"page_number":{"type":"integer","title":"Page Number"},"text":{"type":"string","title":"Text"},"score":{"type":"number","title":"Score"}},"type":"object","required":["document_id","document_title","page_number","text","score"],"title":"CitationInfo"},"DocumentOut":{"properties":{"id":{"type":"integer","title":"Id"},"title":{"type":"string","title":"Title"},"source_filename":{"type":"string","title":"Source Filename"},"upload_date":{"type":"string","format":"date-time","title":"Upload Date"},"version":{"type":"integer","title":"Version"},"status":{"type":"string","title":"Status"}},"type":"object","required":["id","title","source_filename","upload_date","version","status"],"title":"DocumentOut"},"FeedbackOut":{"properties":{"id":{"type":"integer","title":"Id"},"query_id":{"type":"integer","title":"Query Id"},"score":{"type":"integer","title":"Score"},"comment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Comment"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"}},"type":"object","required":["id","query_id","score","comment","timestamp"],"title":"FeedbackOut"},"FeedbackRequest":{"properties":{"query_id":{"type":"integer","title":"Query Id"},"score":{"type":"integer","title":"Score"},"comment":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Comment"}},"type":"object","required":["query_id","score"],"title":"FeedbackRequest"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"QueryHistoryOut":{"properties":{"id":{"type":"integer","title":"Id"},"question":{"type":"string","title":"Question"},"response":{"type":"string","title":"Response"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp"},"feedbacks":{"items":{"$ref":"#/components/schemas/FeedbackOut"},"type":"array","title":"Feedbacks","default":[]}},"type":"object","required":["id","question","response","timestamp"],"title":"QueryHistoryOut"},"QueryRequest":{"properties":{"question":{"type":"string","title":"Question"},"top_k":{"type":"integer","title":"Top K","default":5}},"type":"object","required":["question"],"title":"QueryRequest"},"QueryResponse":{"properties":{"query_id":{"type":"integer","title":"Query Id"},"answer":{"type":"string","title":"Answer"},"citations":{"items":{"$ref":"#/components/schemas/CitationInfo"},"type":"array","title":"Citations"}},"type":"object","required":["query_id","answer","citations"],"title":"QueryResponse"},"Token":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"}},"type":"object","required":["access_token","token_type"],"title":"Token"},"UserOut":{"properties":{"id":{"type":"integer","title":"Id"},"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"role":{"type":"string","title":"Role"}},"type":"object","required":["id","name","email","role"],"title":"UserOut"},"UserRegister":{"properties":{"name":{"type":"string","title":"Name"},"email":{"type":"string","format":"email","title":"Email"},"password":{"type":"string","title":"Password"},"role":{"type":"string","title":"Role","default":"doctor"}},"type":"object","required":["name","email","password"],"title":"UserRegister"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"OAuth2PasswordBearer":{"type":"oauth2","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/login"}}}}}}